Updated Mostly documentation- READ DESCRIPTION
Removed MISC. stuff that was not needed.
Updated documentation and comments for some extra clarity.
Added comments and not to self's (They are denotes towards myself).
Fixed the literally stupiddest grossest 1 line bug, it was horrible.

The loaded cursors in listview are now clickable (this was fixed after the bug was fixed) and are deletable as well.