Read Description***
Refactored the button formatting in the main menu, it looks pretty good now.

Tried fixing the recipe edit page, but I cannot get it right, I will talk to carolyn about helping with that as she is better at the XML stuffs.

Refactored some minor logic to separate it from the GUI.